Introduce new option: KEEPREVERTED.
New option KEEPREVERTED with switches -k/-K to enable and disable
(default).

Feature will stop prevent updating if the latest version has been
downloaded previously (and the checksum passes) but is not the currently
installed version.  This will allow one to revert by re-installing the
previous version (in the case of a regression in plex) and not have it
clobbered the next time plexupdate runs; however, when a new version is
released, it should continue to update from that version and beyond.